+++ src/main/java/absolutelyaya/ultracraft/client/sound/ModularLevelMusic.java
@@ -5,16 +5,17 @@ import net.minecraft.util.Identifier;
 
 public class ModularLevelMusic
 {
-	final Identifier calmID, combatID;
+	final Identifier calmID, combatID, introID;
 	final int combatThreshold;
 	final boolean noCalmdown;
 	
-	public ModularLevelMusic(Identifier calmID, Identifier combatID, int combatThreshold, boolean noCalmdown)
+	public ModularLevelMusic(Identifier calmID, Identifier combatID, int combatThreshold, boolean noCalmdown, Identifier intoID)
 	{
 		this.calmID = calmID;
 		this.combatID = combatID;
 		this.combatThreshold = combatThreshold;
 		this.noCalmdown = noCalmdown;
+		this.introID = intoID;
 	}
 	
 	public SoundEvent getCalmSound()
@@ -29,6 +30,18 @@ public class ModularLevelMusic
 		if(combatID == null)
 			return null;
 		return SoundEvent.of(combatID);
+	}
+	
+	public SoundEvent getIntroSound()
+	{
+		if(introID == null)
+			return null;
+		return SoundEvent.of(introID);
+	}
+	
+	public boolean isHasIntro()
+	{
+		return introID != null;
 	}
 	
 	public int getCombatThreshold()